# esto-laravel-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ESTO Plateforme Doctorale API",
        "description": "API REST complète de la Plateforme de Gestion Doctorale — École Supérieure de Technologie d'Oujda (UMP)",
        "contact": {
            "email": "admin@ump.ac.ma"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur de développement local"
        }
    ],
    "paths": {
...
        "/api/subjects": {
            "get": {
                "tags": [
                    "Sujets"
                ],
                "summary": "Lister tous les sujets de thèse",
                "operationId": "3ec7006ef982f4e44f6dc65d0e55e4bd",
                "parameters": [
                    {
                        "name": "lab",
                        "in": "query",
                        "description": "Filtrer par ID de laboratoire",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des sujets de thèse"
                    }
                }
            },
            "post": {
                "tags": [
                    "Sujets"
                ],
                "summary": "Créer un nouveau sujet de thèse",
                "operationId": "f29ce71563e5fe693862e9070c8ddc4c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titre",
                                    "laboratoire_id"
                                ],
                                "properties": {
                                    "titre": {
                                        "type": "string",
                                        "example": "Intelligence artificielle et santé"
                                    },
                                    "objectif": {
                                        "type": "string",
                                        "example": "Objectif du sujet de recherche"
                                    },
                                    "laboratoire_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "equipe_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "statut": {
                                        "type": "string",
                                        "enum": [
                                            "disponible",
                                            "attribue",
                                            "cloture"
                                        ],
                                        "example": "disponible"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sujet créé avec succès"
                    },
                    "422": {
                        "description": "Données invalides"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/subjects/{id}": {
            "put": {
                "tags": [
                    "Sujets"
                ],
                "summary": "Modifier un sujet de thèse",
                "operationId": "841edf615d9fd5be9671c75caacf3a25",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titre": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "statut": {
                                        "type": "string",
                                        "enum": [
                                            "disponible",
                                            "attribue",
                                            "cloture"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sujet mis à jour"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Sujets"
                ],
                "summary": "Supprimer un sujet de thèse",
                "operationId": "2ad6b7d5f798b22037ff1f0c396640d1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sujet supprimé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}